lsp: Introduce `ServerCompletionConfig`
This is a namespace of within which to place configuration options
which affect the behavior of completions.

The first option introduced is `preferred_insert_behavior` which gives
a hint to the server whether or not the user prefers completions to
`replace` existing text or to only `insert` new text.
