+++ lib/esbonio/changes/531.api.rst
@@ -1,0 +1,1 @@
+``CompletionContext`` objects now expose a ``config`` field that contains any user supplied configuration values affecting completions.
